requirements.txt
streamlit==1.26.0
openai==0.28.1
anthropic==0.3.0
deepseek==1.0.0
qwen==1.0.0
llama-index==0.9.0
gemini==1.0.0
grok==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
markdown==3.4.4
huggingface_hub==0.19.4
PyPDF2==3.0.1
